pyproject.toml
[tool.poetry]
name = "avalanche"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pendulum = "^2.1.2"
google-cloud = "0.34.0"
google-cloud-core = "2.3.2"
google-cloud-storage = "2.7.0"
attrs = "^22.2.0"
cachetools = "^5.3.0"
certifi = "^2022.12.7"
charset-normalizer = "^3.0.1"
click = "^8.1.3"
click-plugins = "^1.1.1"
cligj = "^0.7.2"
Fiona = "^1.9.1"
geopandas = "^0.12.2"
google-api-core = "^2.11.0"
google-auth = "^2.16.2"
google-crc32c = "^1.5.0"
google-resumable-media = "^2.4.1"
googleapis-common-protos = "^1.58.0"
idna = "^3.4"
lxml = "^4.9.2"
munch = "^2.5.0"
numpy = "^1.24.2"
packaging = "^23.0"
pandas = "^1.5.3"
protobuf = "^4.22.0"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.8"
pyproj = "^3.4.1"
python-dateutil = "^2.8.2"
pytz = "^2022.7.1"
requests = "^2.28.2"
rsa = "^4.9"
shapely = "^2.0.1"
six = "^1.16.0"
urllib3 = "^1.26.14"
owslib = "0.28.1"
test-utils = "0.1.0"
geopy = "^2.2.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
